Created initial stencil for project
This is a standard CMake project that builds a shared library and one or
more applications. I created some simple classes for the client and
server, and two executable apps (client.cpp and server.cpp) that
instantiate them. I also copied in an OpenSSL wrapper library that I
wrote for another project, to ensure the dependency on OpenSSL worked,
although it's not currently being used by the client and server.