Added deployment configuration, added checkin request messages
I implemented more of the basic client-server interaction messages,
namely the check-in request that a client sends to the check-in service,
and wrote asynchronous socket code for the client to send the sequence
of messages it needs to send to complete one check-in. I also redesigned
the existing VoterIDRequest messages slightly to make them easier to
sign and verify.

I also created a bunch of configuration files that could be used to set
up a test deployment within multiple terminals of one computer, and put
them in a folder within apps/, which I think is the best home for them
right now.