Fixed bugs in client-server messaging
I neglected to have the servers send the size of the message at the
beginning of the buffer for some of their responses to the client. There
also still seems to be a bug in sending the signature back from the
check-in server (but not the ID server), but I haven't figured it out
yet and I need to switch computers.