Added basic README and clarified role of ID service
After some discussion with others, I realized that the ID service needs
to link identification-document images with unique voter ID numbers that
are used by this system (and presumably the voter registration database)
to identify voters internally. Of course, the mock-up ID service can't
do that, so I made the client tell the ID service which unique ID a
voter should get by placing that number in the first 4 bytes of the "ID
document data" (which is otherwise ignored).