Added voter-list file and voter state table
The check-in service now loads a CSV file containing a list of voters
indexed by UID, saves the UIDs in a std::map in memory, and uses the map
to record which voters have checked in. This is a basic mock-up of the
"registered voter table" that would probably reside in a database in
real life.