CMake fixes, increased CMake version to 3.19
With CMake 3.19 and higher we can specify a version range for
dependencies, not just a minimum version. This allows me to constrain
spdlog to not use version 1.10+ because it is not backward compatible
with spdlog 1.9.2 and below. Also, mutils now supports modern CMake
packaging with export targets, and I learned some improvements to the
Config.cmake file.