Rewrote generate_keys.sh
Made major changes to generate_keys.sh, made it now generate a certificate authority that is then used to sign all the certificates of the clients and servers. Also added all the names of the certificates into config.hpp and config.cpp. Also edited the file extensions to all use .pem for convenience