Updated spdlog dependency to version 1.11
Although Ubuntu's package manager is still stuck on spdlog 1.9.2 and
fmtlib 8.1, most other distros have moved on to the next version of
these libraries (1.11 and 9.1), which have a breaking change in the API
for printing out user-defined types. Changing to the newer version of
spdlog, which maintains backwards compatibility through its next several
versions, will make it easier for Vignesh to work on the code on
non-Ubuntu machines.