Initial attempt at setting up a way to test within Docker
This hasn't been tested yet, but I need to save my work. The intention
is to use docker-compose to set up a Docker network of 6 containers, 2
for the servers and 2 of each client, that each have their own copy of
the source code for this project and can recompile it any time a file
changes without needing to rebuild and relaunch the whole container.